put the two tunnel tests in series so they won't fight
