+++ rust/bin-shared/src/tun_device_manager.rs
@@ -14,12 +14,25 @@ pub use windows as platform;
 pub use platform::TunDeviceManager;
 
 #[cfg(test)]
+#[cfg(any(target_os = "linux", target_os = "windows"))]
 mod tests {
     use super::*;
+    use tracing_subscriber::EnvFilter;
 
     #[tokio::test]
-    #[ignore = "needs sudo"]
-    #[cfg(any(target_os = "linux", target_os = "windows"))]
+    #[ignore = "Needs admin / sudo"]
+    async fn tunnel() {
+        let _ = tracing_subscriber::fmt()
+            .with_env_filter(EnvFilter::from_default_env())
+            .with_test_writer()
+            .try_init();
+
+        // Run these tests in series since they would fight over the tunnel interface
+        // if they ran concurrently
+        create_tun().await;
+        tunnel_drop();
+    }
+
     async fn create_tun() {
         #[cfg(target_os = "windows")]
         {
@@ -35,5 +48,14 @@ mod tests {
 
         let mut tun_device_manager = TunDeviceManager::new().unwrap();
         let _tun = tun_device_manager.make_tun().unwrap();
+    }
+
+    /// Checks for regressions in issue #4765, un-initializing Wintun
+    /// Redundant but harmless on Linux.
+    fn tunnel_drop() {
+        // Each cycle takes about half a second, so this will take a fair bit to run.
+        for _ in 0..50 {
+            let _tun = platform::Tun::new().unwrap(); // This will panic if we don't correctly clean-up the wintun interface.
+        }
     }
 }
